+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The minor version will be incremented upon a breaking change and the patch versi
 - lang: Add `#[error]` attribute to `declare_program!` ([#3757](https://github.com/coral-xyz/anchor/pull/3757)).
 - cli: Replace `anchor verify` to use `solana-verify` under the hood, adding automatic installation via AVM, local path support, and future-proof argument passing ([#3768](https://github.com/solana-foundation/anchor/pull/3768)).
 
+### Removed
+
+- cli: Remove deprecated `anchor publish` command and related code as it's no longer maintained
+
 ### Fixes
 
 - docker: Upgrade `node` to 20.18.0 LTS ([#3687](https://github.com/solana-foundation/anchor/pull/3687)).